+++ Project/Assets/Sparta/Extensions/Multiplayer/ReceivedNetworkMessage.cs
@@ -6,10 +6,10 @@ namespace SocialPoint.Multiplayer
     public class ReceivedNetworkMessage
     {
         public byte MessageType{ get; private set; }
-        public int ChannelId{ get; private set; }
+        public byte ChannelId{ get; private set; }
         public IReader Reader{ get; private set; }
 
-        public ReceivedNetworkMessage(byte type, int chanId, IReader reader)
+        public ReceivedNetworkMessage(byte type, byte chanId, IReader reader)
         {
             MessageType = type;
             ChannelId = chanId;
