+++ Project/Assets/Sparta/Extensions/Multiplayer/UNet/UnetNetworkClient.cs
@@ -106,7 +106,7 @@ namespace SocialPoint.Multiplayer
         {
             var data = new NetworkMessageData {
                 MessageType = UnetNetworkMessage.ConvertType(umsg.msgType),
-                ChannelId = umsg.channelId,
+                ChannelId = (byte)umsg.channelId,
             };                
             if(_receiver != null)
             {
